pyproject.toml
[tool.poetry]
name = "energy-forecasting"
version = "0.1.0"
description = ""
authors = [
"Paul Iusztin <paul.lusztin@coreai.ai>",
"Kurtis Pykes <kurtispykes@gmail.com>" 
]

[tool.poetry.dependencies]
python = "~3.9"
pyarrow = "^11.0.0"
tqdm = "^4.64.1"
category-encoders = "^2.6.0"
wandb = "^0.13.10"
matplotlib = "^3.7.0"
hopsworks = "^3.0.5"
python-dotenv = "^0.21.1"
lightgbm = "^3.3.5"
sktime = "^0.16.1"
seaborn = "^0.12.2"
pmdarima = "^2.0.2"
google-cloud-storage = "^2.7.0"
yarl = "^1.8.2"
fire = "^0.5.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
